requirement.txt
aiohttp==3.8.5
aiosignal==1.3.1
altgraph==0.17.3
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
asgiref==3.7.2
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
auto-py-to-exe==2.36.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
blis==0.7.11
bottle==0.12.25
bottle-websocket==0.2.9
Brotli==1.1.0
catalogue==2.0.10
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
confection==0.1.3
consonance==0.1.2
cryptography==41.0.3
cymem==2.0.8
dataclasses-json==0.6.1
debugpy==1.8.0
decorator==5.1.1
diskcache==5.6.3
dissononce==0.34.3
distlib==0.3.7
Django==4.2.3
EasyProcess==1.1
Eel==0.16.0
entrypoint2==1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.2
executing==2.0.0
filelock==3.12.3
Flask==2.3.2
frozenlist==1.4.0
fsspec==2023.9.2
future==0.18.3
gevent==23.7.0
gevent-websocket==0.10.1
greenlet==2.0.2
h11==0.14.0
huggingface-hub==0.17.3
idna==3.4
imutils==0.5.4
ipykernel==6.25.2
ipython==8.16.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.3.1
jupyter_core==5.3.2
kodekloud-downloader==0.1.2
langchain==0.0.304
langcodes==3.3.0
langsmith==0.0.41
llama-index==0.8.37
llama_cpp_python==0.2.11
markdownify==0.11.6
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
multidict==6.0.4
murmurhash==1.0.10
mutagen==1.47.0
mypy-extensions==1.0.0
mysql-connector-python==8.1.0
nest-asyncio==1.5.8
networkx==3.1
nltk==3.8.1
numexpr==2.8.7
numpy==1.25.2
openai==0.28.1
opencv-python==4.8.0.76
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathy==0.10.2
pefile==2023.2.7
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
preshed==3.0.9
prettytable==3.9.0
prompt-toolkit==3.0.39
protobuf==4.21.12
psutil==5.9.5
pure-eval==0.2.2
PyAutoGUI==0.9.54
pycparser==2.21
pycryptodomex==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
PyGetWindow==0.0.9
Pygments==2.16.1
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6
PyMsgBox==1.0.9
pyodbc==4.0.39
pyparsing==3.1.1
pypdf==3.16.2
pyperclip==1.8.2
pyqt-llamaindex==0.0.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
python-axolotl==0.2.2
python-axolotl-curve25519==0.4.1.post2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytweening==1.0.7
pytz==2023.3
pywhatkit==5.4
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
schedule==1.2.0
scikit-learn==1.3.1
scipy==1.11.3
selenium==4.10.0
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.21
sqlparse==0.4.4
srsly==2.4.8
stack-data==0.6.3
sympy==1.12
tenacity==8.2.3
thinc==8.1.12
threadpoolctl==3.2.0
tiktoken==0.5.1
tokenizers==0.13.3
torch==2.0.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.1
transformers==4.33.3
transitions==0.9.0
tree-sitter==0.20.2
tree-sitter-languages==1.7.0
trio==0.22.2
trio-websocket==0.10.3
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
validators==0.20.0
virtualenv==20.24.4
wasabi==1.1.2
wcwidth==0.2.6
webdriver-manager==3.9.1
websockets==11.0.3
Werkzeug==2.3.7
whichcraft==0.6.1
wikipedia==1.4.0
wsproto==1.2.0
yarl==1.9.2
yowsup2==3.2.3
yt-dlp==2023.9.24
zope.event==5.0
zope.interface==6.0
